+++ src/repositories/ChannelRepository.js
@@ -189,100 +189,6 @@ export const findChatLogById = async (id) => {
     }
 };
 
-export const createChat = async (id, channelId, content) => {
-    try {
-        return await prisma.chatMessage.create({
-            data: {
-                channelId,
-                sendUserId: id,
-                content,
-                createdAt: dbNow(),
-            },
-            include: {
-                sendUser: {
-                    select: {
-                        id: true,
-                        email: true,
-                        nickname: true,
-                        profile: {
-                            select: {
-                                profileImage: {
-                                    select: {
-                                        src: true,
-                                    },
-                                },
-                            },
-                        },
-                    },
-                },
-            },
-        });
-    } catch (err) {
-        console.error(err);
-    }
-};
-export const findLastId = async (channelId, lastId) => {
-    try {
-        return await prisma.chatMessage.findMany({
-            where: {
-                AND: [
-                    {
-                        channelId,
-                    },
-                    {
-                        id: {
-                            lt: lastId,
-                        },
-                    },
-                ],
-            },
-            orderBy: {
-                id: "desc",
-            },
-        });
-    } catch (err) {
-        console.error(err);
-    }
-};
-export const findChatByChannelId = async (channelId, lastId, limit) => {
-    try {
-        return await prisma.chatMessage.findMany({
-            take: limit,
-            where: {
-                channelId,
-            },
-            cursor: lastId
-                ? {
-                      id: lastId,
-                  }
-                : undefined,
-            orderBy: {
-                createdAt: "desc",
-            },
-            include: {
-                sendUser: {
-                    select: {
-                        id: true,
-                        email: true,
-                        nickname: true,
-                        profile: {
-                            select: {
-                                profileImage: {
-                                    select: {
-                                        src: true,
-                                    },
-                                },
-                            },
-                        },
-                    },
-                },
-            },
-        });
-    } catch (err) {
-        console.error(err);
-    }
-};
-
 export const findByKeyword = async (category, keyword, offset = 0) => {
     try {
         return await prisma.channel.findMany({
