+++ src/repositories/ChannelRoomRepository.js
@@ -220,181 +220,6 @@ export const findParticipantRoom = async (id) => {
     }
 };
 
-export const findChatLastId = async (roomId, lastId) => {
-    try {
-        return await prisma.chatMessage.findMany({
-            where: {
-                AND: [
-                    {
-                        channelRoomId: roomId,
-                    },
-                    {
-                        id: {
-                            lt: lastId,
-                        },
-                    },
-                ],
-            },
-            orderBy: {
-                id: "desc",
-            },
-        });
-    } catch (err) {
-        console.error(err);
-    }
-};
-
-export const findChatByRoomId = async (channelRoomId, lastId, limit) => {
-    try {
-        return await prisma.chatMessage.findMany({
-            take: limit,
-            where: {
-                channelRoomId,
-            },
-            cursor: lastId
-                ? {
-                      id: lastId,
-                  }
-                : undefined,
-            orderBy: {
-                createdAt: "desc",
-            },
-            include: {
-                answeredMessage: {
-                    include: {
-                        sendUser: {
-                            select: {
-                                id: true,
-                                email: true,
-                                nickname: true,
-                                profile: {
-                                    select: {
-                                        profileImage: {
-                                            select: {
-                                                src: true,
-                                            },
-                                        },
-                                    },
-                                },
-                            },
-                        },
-                    },
-                },
-                sendUser: {
-                    select: {
-                        id: true,
-                        email: true,
-                        nickname: true,
-                        profile: {
-                            select: {
-                                profileImage: {
-                                    select: {
-                                        src: true,
-                                    },
-                                },
-                            },
-                        },
-                    },
-                },
-            },
-        });
-    } catch (err) {
-        console.error(err);
-    }
-};
-
-export const createChat = async (id, channelRoomId, content) => {
-    try {
-        return await prisma.chatMessage.create({
-            data: {
-                channelRoomId,
-                sendUserId: id,
-                content,
-                createdAt: dbNow(),
-            },
-            include: {
-                sendUser: {
-                    select: {
-                        id: true,
-                        email: true,
-                        nickname: true,
-                        profile: {
-                            select: {
-                                profileImage: {
-                                    select: {
-                                        src: true,
-                                    },
-                                },
-                            },
-                        },
-                    },
-                },
-            },
-        });
-    } catch (err) {
-        console.error(err);
-    }
-};
-
-export const createChatAnswer = async (
-    id,
-    channelRoomId,
-    content,
-    answeredId
-) => {
-    try {
-        return await prisma.chatMessage.create({
-            data: {
-                channelRoomId,
-                sendUserId: id,
-                content,
-                createdAt: dbNow(),
-                answeredId,
-            },
-            include: {
-                answeredMessage: {
-                    include: {
-                        sendUser: {
-                            select: {
-                                id: true,
-                                email: true,
-                                nickname: true,
-                                profile: {
-                                    select: {
-                                        profileImage: {
-                                            select: {
-                                                src: true,
-                                            },
-                                        },
-                                    },
-                                },
-                            },
-                        },
-                    },
-                },
-                sendUser: {
-                    select: {
-                        id: true,
-                        email: true,
-                        nickname: true,
-                        profile: {
-                            select: {
-                                profileImage: {
-                                    select: {
-                                        src: true,
-                                    },
-                                },
-                            },
-                        },
-                    },
-                },
-            },
-        });
-    } catch (err) {
-        console.error(err);
-    }
-};
-
 export const updateCloseRoom = async (id) => {
     try {
         return await prisma.channelRoom.update({
